+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  - Updated Some Client Page Packages
  - Fixed mangapark base url
  - Styled Change User Title Text Box
+ - Support for Bato
 
 ## Change Log 3.1.8
  - MangaPark Support